FIX: Fix the sourcing the {HOME}/.bashrc in docker
The autogenerated bashrc file has a guard system to prevent
sourcing from non-interactive shells.
Since we need to test the installation scripts in a simulated environment
as close as possible to the real one, we need to source the bashrc file
during test phase.

To do this, we need to enable interactive mode in the docker container.
NOTE: This is not a interactive docker session (--it flag).
In addition to simulate the user session, the login shell option has been added to the docker recipe.

The docker now has the following options:
-i (interactive) source the bashrc file (sets PS1, PS2, etc.)
-l (login) source the profile file

NOTE: The interactive option produces the following warnings:
    bash: cannot set terminal process group (-1): Inappropriate ioctl for device
    bash: no job control in this shell
The warnings are harmless and can be ignored