Github Action CI setup for Ubuntu
This commit sets up a GitHub Action workflow for end-to-end testing on Ubuntu.
From now, we will support several Ubuntu LTS versions explicity:
    - Ubuntu 20.04 (Focal Fossa)
    - Ubuntu 22.04 (Jammy Jellyfish)
    - Ubuntu 24.04 (Noble Narwhal)
    - Ubuntu 25.04 (Plucky Puffin)

The main feature of the setup is the ability to run podman containers
under the docker. This will simulate the real environment, where the
podman will be used to run the containers on real systems.

To run the docker containers with ability to run podman containers,
the following command is used:
    docker run --privileged --security-opt seccomp=unconfined

IMPORTANT: The test setup is created as a template. Other distributions
           should follow the sam pattern.

The pattern is as follows:
    - as ROOT install common packages (fonts, bash, locales, etc.)
    - as ROOT make basic configurations (e.g. set locale)
    - as ROOT add the new user (with no sudo access)
    - as ROOT install podman and all required packages
        + grant the user a subuid and subgid
        + set XDG_RUNTIME_DIR
        + set uid for newuidmap
        + update ca-certificates
        + remount '/' with rshare option
    - as USER execute the project to make tests

Signed-off-by: Michal Kukowski <kukossw@gmail.com>