Github Action CI setup for RHEL
This commit sets up a GitHub Action workflow for end-to-end testing on RHEL.
From now, we will support several RedHat (RHEL) versions explicity:
    - RHEL 8
    - RHEL 9
    - RHEL 10

The template from Ubuntu has been used as a base.
The main difference is that the RHEL but default requires subscription
to access the repositories. To mock the subscription, we use
the epel mirror repositories (mainly from fedora project).

To run the docker containers with ability to run podman containers,
the following command is used:
    docker run --privileged --security-opt seccomp=unconfined

Signed-off-by: Michal Kukowski <kukossw@gmail.com>