Github Action CI setup for RockyLinux
This commit sets up a GitHub Action workflow for end-to-end testing on RockyLinux.
From now, we will support several RockyLinux versions explicity:
    - RockyLinux 8
    - RockyLinux 9

NOTE: The RockyLinux 10 is already available as the image, but dockerhub
      does not have the image available yet. As soon as it is available,
      the explicit support for RockyLinux 10 will be added.

Since the RockyLinux is based on the RHEL, the Dockerfiles are based on the
RHEL Dockerfiles. The only differences are:
    - The workaround for paid RHEL subscription is removed. Since the
      RockyLinux is free, there is no need to use the workaround.
    - The `sudo mount --make-rshared /` has been not working.
      Instead of the `unshare -m --propagation shared` command is used as the workaround.

To run the docker containers with ability to run podman containers,
the following command is used:
    docker run --privileged --security-opt seccomp=unconfined

Signed-off-by: Michal Kukowski <kukossw@gmail.com>