Github Action CI setup for Fedora
This commit sets up a GitHub Action workflow for end-to-end testing on Fedora.
From now, we will support several Fedora versions explicity:
    - Fedora 38
    - Fedora 39
    - Fedora 40
    - Fedora 41
    - Fedora 42

Since Fedora and free RHELs (RockyLinux is a good example) are very similar,
the Dockerfles are based on RockyLinux.
Fedora 38 and 39 have the same issues with default podman config
(cgroup and systemd journal) as Debian 10 and 11.
The same workaround has been applied.

To run the docker containers with ability to run podman containers,
the following command is used:
    docker run --privileged --security-opt seccomp=unconfined

Signed-off-by: Michal Kukowski <kukossw@gmail.com>