Github Action CI setup for SUSE
This commit sets up a GitHub Action workflow for end-to-end testing on SUSE.
From now, we will support several SUSE versions explicity:
    - SUSE 15 (SLE 15)

SUSE is a Linux distribution that is widely used in enterprise environments,
the major versions are released every 3-5 years. SUSE 12 reached its end of life in 2021,
that's why we are going to support only SUSE 15 for now.

To run the docker containers with ability to run podman containers,
the following command is used:

docker run --privileged --security-opt seccomp=unconfined

Signed-off-by: Michal Kukowski <kukossw@gmail.com>