Fix CI connection error flakiness
Sometimes on Github Action CI the package manager
has connection error. The error is not connected with our settings
nor the code. The issue is the internal Github Action issue and we
cannot do anything about that.

To reduce flakiness the retry:10 option has been added to following
package managers:
    - apt (Ubuntu + Debian)
    - dnf (RHEL + Fedora + CentOS + CentOSStream + Rocky)

The only missing one is the zypper (SUSE + OpenSUSE), where
the retry option is not available.

Signed-off-by: Michal Kukowski <kukossw@gmail.com>